+++ src/lib/utils/ValidatingFileRegistry.ts
@@ -1,16 +1,19 @@
-import { FileRegistry } from "../models/FileRegistry.js";
-import { isFile } from "./fs.js";
+import { type FileId, FileRegistry } from "../models/FileRegistry.js";
 import type { Deserializer, JSONOutput } from "#serialization";
 import { i18n, type NormalizedPath, NormalizedPathUtils } from "#utils";
+import { existsSync } from "fs";
 
 export class ValidatingFileRegistry extends FileRegistry {
     override register(
         sourcePath: NormalizedPath,
         relativePath: NormalizedPath,
-    ): { target: number; anchor: string | undefined } | undefined {
+    ): { target: FileId; anchor: string | undefined } | undefined {
         const absolute = NormalizedPathUtils.resolve(NormalizedPathUtils.dirname(sourcePath), relativePath);
         const absoluteWithoutAnchor = absolute.replace(/#.*/, "");
-        if (!isFile(absoluteWithoutAnchor)) {
+        // Note: We allow paths to directories to be registered here, but the AssetsPlugin will not
+        // copy them to the output path. This is so that we can link to directories and associate them
+        // with reflections in packages mode.
+        if (!existsSync(absoluteWithoutAnchor)) {
             return;
         }
         return this.registerAbsolute(absolute);
@@ -19,9 +22,9 @@ export class ValidatingFileRegistry extends FileRegistry {
     override fromObject(de: Deserializer, obj: JSONOutput.FileRegistry) {
         for (const [key, val] of Object.entries(obj.entries)) {
             const absolute = NormalizedPathUtils.resolve(de.projectRoot, val);
-            if (!isFile(absolute)) {
+            if (!existsSync(absolute)) {
                 de.logger.warn(
-                    i18n.saved_relative_path_0_resolved_from_1_is_not_a_file(
+                    i18n.saved_relative_path_0_resolved_from_1_does_not_exist(
                         val,
                         de.projectRoot,
                     ),
@@ -29,7 +32,7 @@ export class ValidatingFileRegistry extends FileRegistry {
                 continue;
             }
 
-            de.oldFileIdToNewFileId[+key] = this.registerAbsolute(absolute).target;
+            de.oldFileIdToNewFileId[+key as FileId] = this.registerAbsolute(absolute).target;
         }
 
         de.defer((project) => {
@@ -39,7 +42,7 @@ export class ValidatingFileRegistry extends FileRegistry {
                 );
                 if (refl) {
                     this.mediaToReflection.set(
-                        de.oldFileIdToNewFileId[+media]!,
+                        de.oldFileIdToNewFileId[+media as FileId]!,
                         refl.id,
                     );
                 } else {