+++ src/test/converter.test.ts
@@ -21,7 +21,6 @@ import type { ModelToObject } from "../lib/serialization/schema.js";
 import { getExpandedEntryPointsForPaths, normalizePath } from "../lib/utils/index.js";
 import { getConverterApp, getConverterBase, getConverterProgram } from "./programs.js";
 import { FileRegistry } from "../lib/models/FileRegistry.js";
-import { ValidatingFileRegistry } from "../lib/utils/ValidatingFileRegistry.js";
 
 const comparisonSerializer = new Serializer();
 comparisonSerializer.addSerializer({
@@ -201,7 +200,7 @@ describe("Converter", function () {
                 it(`[${file}] converts fixtures`, function () {
                     before();
                     resetReflectionID();
-                    app.files = new ValidatingFileRegistry();
+                    app.files = new FileRegistry();
                     const entryPoints = getExpandedEntryPointsForPaths(
                         app.logger,
                         [path],